+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## [0.0.1] - 2022.07.03
+- loadGame function added
 - saveGame function added